Refactored the packages of the backend module
