+++ equinox-backend/src/main/java/com/tecknobit/equinoxbackend/apis/batch/JoinTableSyncBatchItem.java
@@ -1,4 +1,4 @@
-package com.tecknobit.equinoxbackend.batch;
+package com.tecknobit.equinoxbackend.apis.batch;
 
 import com.tecknobit.equinoxbackend.annotations.BatchQueryItem;
 import com.tecknobit.equinoxbackend.environment.services.builtin.service.EquinoxItemsHelper;
@@ -16,7 +16,6 @@ import java.util.List;
  * @author N7ghtm4r3 - Tecknobit
  * @since 1.0.8
  */
-@Deprecated(forRemoval = true, since = "MOVED INTO apis.batch PACKAGE")
 @BatchQueryItem
 public class JoinTableSyncBatchItem<O, D> implements EquinoxItemsHelper.ComplexBatchItem {
