+++ equinox-backend/src/main/java/com/tecknobit/equinoxbackend/apis/events/EquinoxEventsEmitter.java
@@ -1,4 +1,4 @@
-package com.tecknobit.equinoxbackend.events;
+package com.tecknobit.equinoxbackend.apis.events;
 
 import jdk.jfr.Experimental;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,7 +14,6 @@ import org.springframework.stereotype.Component;
  * @author N7ghtm4r3 - Tecknobit
  * @since 1.1.1
  */
-@Deprecated(forRemoval = true, since = "MOVED INTO apis.events PACKAGE")
 @Component
 @Experimental
 public abstract class EquinoxEventsEmitter<T extends Enum<?>> {
